Add separate module for creating functionality
